requirements.txt
boto3==1.28.7
pymongo==4.5.0
python-dotenv==1.0.0